requirements.txt
attrs==22.2.0
aws-sam-translator==1.63.0
aws-xray-sdk==2.11.0
boto3==1.26.102
botocore==1.29.102
certifi==2022.12.7
cffi==1.15.1
cfn-lint==0.76.1
charset-normalizer==3.1.0
cryptography==40.0.1
docker==6.0.1
ecdsa==0.18.0
graphql-core==3.2.3
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
jschema-to-python==1.2.3
jsondiff==2.0.0
jsonpatch==1.32
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
jsonschema-spec==0.1.4
junit-xml==1.9
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
moto==4.1.6
mpmath==1.3.0
networkx==3.0
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.6
packaging==23.0
pathable==0.4.3
pbr==5.11.1
pkgutil-resolve-name==1.3.10
py-partiql-parser==0.1.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.2
responses==0.23.1
rfc3339-validator==0.1.4
rsa==4.9
s3transfer==0.6.0
sarif-om==1.0.4
six==1.16.0
sshpubkeys==3.3.1
sympy==1.11.1
types-PyYAML==6.0.12.9
typing-extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
xmltodict==0.13.0
zipp==3.15.0
